examples: xlnx: linux: Update demo to match latest
As part of this, move the latest logic that was present in
examples/system/linux/zynqmp to examples/system/linux/xlnx.
