+++ examples/system/linux/CMakeLists.txt
@@ -1,4 +1,3 @@
-
-if (EXISTS ${CMAKE_CURRENT_SOURCE_DIR}/${PROJECT_MACHINE})
-  add_subdirectory(${PROJECT_MACHINE})
-endif (EXISTS ${CMAKE_CURRENT_SOURCE_DIR}/${PROJECT_MACHINE})
+if(EXISTS ${CMAKE_CURRENT_SOURCE_DIR}/${PROJECT_VENDOR})
+    add_subdirectory(${PROJECT_VENDOR})
+endif(EXISTS ${CMAKE_CURRENT_SOURCE_DIR}/${PROJECT_VENDOR})
