--- examples/system/linux/zynqmp/zynqmp_amp_demo/sys_init.c
@@ -1,32 +1,0 @@
-
-/*
- * Copyright (c) 2016, Xilinx Inc. and Contributors. All rights reserved.
- *
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-#include <stdio.h>
-#include <unistd.h>
-#include <metal/sys.h>
-#include "common.h"
-
-int sys_init(void)
-{
-	struct metal_init_params init_param = METAL_INIT_DEFAULTS;
-	int ret;
-
-	ret = metal_init(&init_param);
-	if (ret)
-		LPERROR("Failed to initialize libmetal\n");
-	return ret;
-}
-
-void sys_cleanup(void)
-{
-	metal_finish();
-}
-
-void wait_for_interrupt(void)
-{
-	return;
-}
