--- examples/system/linux/zynqmp/zynqmp_amp_demo/sys_init.h
@@ -1,16 +1,0 @@
-/******************************************************************************
- *
- * Copyright (C) 2017 Xilinx, Inc.  All rights reserved.
- *
- * SPDX-License-Identifier: BSD-3-Clause
- *
- ******************************************************************************/
-
-#ifndef __SYS_INIT_H__
-#define __SYS_INIT_H__
-
-int sys_init(void);
-void sys_cleanup(void);
-
-#endif /* __SYS_INIT_H__ */
-
